// src/app/models/material.model.ts
export class Material{
  material_name: string;
  material_category_id: number;
  gold: number;
  silver: number;
  is_main_production_material: number;
  main_material_id: number;

  constructor(public id: number){}
}

// export class CustomerModel {
//   mobile1: string;
//   mobile2: string;
//   address1: string;
//   address2: string;
//   city: string;
//   state: string;
//   po: string;
//   area: string;
//   pin: string;
//   constructor(public id: number,
//               public person_name: string,
//               public email: string,
//               public customer_category_id: number
//   ) {}
// }